[1.2.0-0BQ11] - Fixed nasty PHP5 issue in ui/web/userMod.php with an even more nasty work around. - Discovered data loss of input from composite form fields on non-active tabs of pages using pagedBlock.
This means:

If you have a page that has multiple tabs (like userMod.php) then usually the variable data from the
non active tab should be in the source code as 'HIDDEN' form fields.

Unfortunately this is NOT the case when you use composite form fields. In that case the data will be
gone!

I have yet to test if this is just due to some weird coding in userMod.php, or if this applies to
all pages that use pagedBlock for tabs and composite form fields.