2010-05-12
In preparation for CentOS 6.0 I created the new 5107R branch for BlueOnyx on CentOS 6.

This is based on the 5106R branch as of today, but additional code will be checked in on it eventually.

This is purely a development and NOT yet ready for production branch until we say otherwise.